뷰 분리, 디테일 뷰모델 업데이트
